♻️ (trace): Extract severity filter constants to dedicated module
