+++ dashboard/src/components/trace/TraceTimelineFilter.tsx
@@ -1,19 +1,6 @@
 import type { KeyboardEvent } from 'react'
-import type { TraceSeverity } from '../../features/trace/types'
+import { ALL_ON, SEVERITY_KEYS, type SeverityFilter, type SeverityKey } from './severityFilter'
 import './TraceTimelineFilter.css'
-
-export type SeverityKey = TraceSeverity | 'neutral'
-
-export const SEVERITY_KEYS: readonly SeverityKey[] = ['critical', 'warning', 'info', 'neutral'] as const
-
-export type SeverityFilter = Readonly<Record<SeverityKey, boolean>>
-
-export const ALL_ON: SeverityFilter = {
-  critical: true,
-  warning: true,
-  info: true,
-  neutral: true,
-}
 
 export interface TraceTimelineFilterProps {
   readonly value: SeverityFilter
